feat : 오른쪽 바 초안
